Configure Project
To begin with, let's create a default single view app project with support for iOS 12 and above.
Configure the app to run only for iPhones in portrait only and set the initial app icon.
Add CocoaPods support for 3rd party dependency management.
Add our first pod Apollo iOS which is a strongly-typed, caching GraphQL client for iOS, written in Swift.