Add Network Clients
Added the 'network (clients) layer' with a configuration struct NetworkConfig.
The configuration holds only one item atm, which is the swapi endpoint.
Added a custom error list via the NetworkError enum. Only generic and noData erros atm.
Created a base NetworkClient class that holds the Apollo Client.
The client also defines a Result type for handling responss and a Paged type for handling paged responses.
For now we have the FilmClient that operates on Film entitites using the GraphQL generated API, as well as
the CharacterClient that operates on Characters/People entity.
These clients are used to convert the responses from GraphQL to usable models or errors within our app.