Add All Films Screen
Update the dir structure of the project. Split it in layer: ViewModel, View (UI).
Add StoryboardPresentable protocol for easy instantiation from storyboards.
Add ViewModelDelegate protocol for all view models to comply with didChange, didFail.
Add an empty (for now) ViewModel protocol to be implemented by all view models.

Update the network clients to return the request, so it can be referenced and cancelled if needed.

Add the first UI screen to display a list of all films.
To support this, there is a FilmListViewModel which ties into FilmListViewController.
A custom cell class is used FilmTableViewCell that can configure itself based on film data.
The views for the controller and cell are in Main.storyboard.
This screen is the initial startup screen and is being setup in didFinishLaunchingWithOptions,
because we need to manually inject the view model.