Add Film Details Screen
Added a factory class for view controllers.
It instantiates a view model and a view controller and does the initial binding.

Add the UI screen to display film details.
To support this, there is a FilmDetailsViewModel which ties into FilmDetailsViewController.
Custom cell classes are used FilmDetailItemTableViewCell, FilmDetailActionTableViewCell
that can configure themselves based on film property data or actions needed to be displayed.
The views for the controller and cell are in Main.storyboard.
This screen is accessable when a user taps on a film item in the film list (first app screen).