Add All Characters Screen
Add the UI screen to display all characters with paging. New data is fetched when the user scrolls to bottom.
To support this, there is a CharacterListViewModel which ties into CharacterListViewController.
Custom cell class is used CharacterTableViewCell that can configure itself based on character data.
The views for the controller and cell are in Main.storyboard.
This screen is accessable when a user taps on 'Characters' action inside film details screen.

Added an extension to UIScrollView that can detect if the scroll view bounces.