Add Loading Cell
Add a LoadingTableViewCell class that will start animating once it moves to the window and stop when it moves out.
Introduce a new enum 'DataType' to be used to distinguish between a placeholder and actual value.
Update the 'All Films' screen to utilize LoadingTableViewCell.
Update the 'Film Details' screen to utilize LoadingTableViewCell.